backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
Pillow==10.2.0
requests==2.31.0
numpy>=1.24.0
scipy==1.11.4
scikit-learn==1.3.2